Revert "change accord to video"
